chore(rsc): Make Router-related imports safer for RSC
Safer, as in not importing React unless absolutely needed. So we don't
end up with having two instances of React imported, or having React not
be initialized properly
