+++ packages/internal/src/routes.ts
@@ -3,7 +3,7 @@ import path from 'path'
 import chalk from 'chalk'
 
 import { getPaths, getRouteHookForPage } from '@redwoodjs/project-config'
-import { getRouteRegexAndParams } from '@redwoodjs/router'
+import { getRouteRegexAndParams } from '@redwoodjs/router/dist/util'
 import { getProject } from '@redwoodjs/structure/dist/index.js'
 import type { RWRoute } from '@redwoodjs/structure/dist/model/RWRoute'
