+++ packages/testing/src/web/MockRouter.tsx
@@ -7,7 +7,8 @@ import type React from 'react'
 // @ts-ignore
 import { isValidRoute } from '@redwoodjs/router/dist/route-validators'
 import type { RouterProps } from '@redwoodjs/router/dist/router'
-import { flattenAll, replaceParams } from '@redwoodjs/router/dist/util'
+import { flattenAll } from '@redwoodjs/router/dist/react-util'
+import { replaceParams } from '@redwoodjs/router/dist/util'
 // @ts-ignore
 export * from '@redwoodjs/router/dist/index'
